web06de20: add segurança nas rotas

+++ src/routers/FuncionarioRouter.py
@@ -8,16 +8,18 @@ from domain.schemas.FuncionarioSchema import (
     FuncionarioUpdate,
     FuncionarioResponse
 )
+from domain.schemas.AuthSchema import FuncionarioAuth
 # Infra
 from infra.orm.FuncionarioModel import FuncionarioDB
 from infra.database import get_db
 from infra.security import get_password_hash
+from infra.dependencies import get_current_active_user, require_group
 
 router = APIRouter()
 
 # Criar as rotas/endpoints: GET, POST, PUT, DELETE
-@router.get("/funcionario/", response_model=List[FuncionarioResponse], tags=["Funcionário"], status_code=status.HTTP_200_OK)
-async def get_funcionario(db: Session = Depends(get_db)):
+@router.get("/funcionario/", response_model=List[FuncionarioResponse], tags=["Funcionário"], status_code=status.HTTP_200_OK, summary="Listar todos os funcionários")
+async def get_funcionario(db: Session = Depends(get_db), current_user: FuncionarioAuth = Depends(require_group([1]))):
     """Retorna todos os funcionários"""
     try:
         funcionarios = db.query(FuncionarioDB).all()
@@ -27,9 +29,9 @@ async def get_funcionario(db: Session = Depends(get_db)):
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail=f"Erro ao buscar funcionários: {str(e)}"
             )
-        
-@router.get("/funcionario/{id}", response_model=FuncionarioResponse, tags=["Funcionário"], status_code=status.HTTP_200_OK)
-async def get_funcionario(id: int, db: Session = Depends(get_db)):
+
+@router.get("/funcionario/{id}", response_model=FuncionarioResponse, tags=["Funcionário"], status_code=status.HTTP_200_OK, summary="Buscar funcionário por ID")
+async def get_funcionario_by_id(id: int, db: Session = Depends(get_db), current_user: FuncionarioAuth = Depends(get_current_active_user)):
     """Retorna um funcionário específico pelo ID"""
     try:
         funcionario = db.query(FuncionarioDB).filter(FuncionarioDB.id == id).first()
@@ -44,8 +46,8 @@ async def get_funcionario(id: int, db: Session = Depends(get_db)):
             detail=f"Erro ao buscar funcionário: {str(e)}"
             )
 
-@router.post("/funcionario/", response_model=FuncionarioResponse, status_code=status.HTTP_201_CREATED, tags=["Funcionário"])
-async def post_funcionario(funcionario_data: FuncionarioCreate, db: Session = Depends(get_db)):
+@router.post("/funcionario/", response_model=FuncionarioResponse, status_code=status.HTTP_201_CREATED, tags=["Funcionário"], summary="Criar novo funcionário")
+async def post_funcionario(funcionario_data: FuncionarioCreate, db: Session = Depends(get_db), current_user: FuncionarioAuth = Depends(require_group([1]))):
     """Cria um novo funcionário"""
     try:
         # Verifica se já existe funcionário com este CPF
@@ -78,8 +80,8 @@ async def post_funcionario(funcionario_data: FuncionarioCreate, db: Session = De
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"Erro ao criar funcionário: {str(e)}"
         )
 
-@router.put("/funcionario/{id}", response_model=FuncionarioResponse, tags=["Funcionário"], status_code=status.HTTP_200_OK)
-async def put_funcionario(id: int, funcionario_data: FuncionarioUpdate, db: Session = Depends(get_db)):
+@router.put("/funcionario/{id}", response_model=FuncionarioResponse, tags=["Funcionário"], status_code=status.HTTP_200_OK, summary="Atualizar funcionário")
+async def put_funcionario(id: int, funcionario_data: FuncionarioUpdate, db: Session = Depends(get_db), current_user: FuncionarioAuth = Depends(require_group([1]))):
     """Atualiza um funcionário existente"""
     try:
         funcionario = db.query(FuncionarioDB).filter(FuncionarioDB.id == id).first()
@@ -112,8 +114,8 @@ async def put_funcionario(id: int, funcionario_data: FuncionarioUpdate, db: Sess
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"Erro ao atualizar funcionário: {str(e)}"
         )
 
-@router.delete("/funcionario/{id}", status_code=status.HTTP_204_NO_CONTENT, tags=["Funcionário"], summary="Remover funcionário")
-async def delete_funcionario(id: int, db: Session = Depends(get_db)):
+@router.delete("/funcionario/{id}", status_code=status.HTTP_204_NO_CONTENT, tags=["Funcionário"], summary="Remover funcionário - apenas ADMIN")
+async def delete_funcionario(id: int, db: Session = Depends(get_db), current_user: FuncionarioAuth = Depends(require_group([1]))):
     """Remove um funcionário"""
     try:
         funcionario = db.query(FuncionarioDB).filter(FuncionarioDB.id == id).first()
@@ -121,6 +123,11 @@ async def delete_funcionario(id: int, db: Session = Depends(get_db)):
             raise HTTPException(
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail="Funcionário não encontrado"
+            )
+        if current_user.id == id:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Não é possível excluir seu próprio usuário"
             )
         db.delete(funcionario)
         db.commit()
